remove duped code
