+++ controllers/index.js
@@ -1,7 +1,6 @@
 import { searchMovie } from "./searchMovie.js";
 import { poster } from "./poster.js";
-import { letterboxdWatchlist } from "./letterboxdWatchlist.js";
-import { letterboxdCustomList } from "./letterboxdCustomList.js";
+import { letterboxdWatchlist, letterboxdCustomList } from "./letterboxdWatchlist.js";
 import { alternativeSearch } from "./alternativeSearch.js";
 import { proxy } from "./proxy.js";
